refactor(style): refactor Header, Footer Component stlyed-components
