+++ .eslintrc.json
@@ -18,6 +18,7 @@
     "no-unused-vars": "off",
     "@typescript-eslint/no-unused-vars": ["error"],
     "@typescript-eslint/no-unnecessary-type-constraint" : "off",
+    "import/prefer-default-export": "off",
     "import/order": [
       "error", 
       {
