+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { AppContainer } from './GlobalStyle';
-import Footer from 'components/Footer';
-import Header from 'components/Header';
+import Footer from 'components/Footer/Footer';
+import Header from 'components/Header/Header';
 import AboutPage from 'pages/AboutPage';
 import ContactPage from 'pages/ContactPage';
 import LandingPage from 'pages/LandingPage';
