+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import FooterLogo from 'assets/footer/FooterLogo.svg';
+import * as S from 'components/Footer/Footer.style';
+
+const Footer = () => {
+  const { t } = useTranslation();
+
+  return (
+    <S.Container>
+      <S.InnerContainer>
+        <FooterLogo />
+        <S.NavContainer>
+          <div>
+            <S.Link href={'mailto:vintzscompany@gmail.com'}>
+              vintzscompany@gmail.com
+            </S.Link>
+          </div>
+          <p>{t('footer.copyright')}</p>
+        </S.NavContainer>
+      </S.InnerContainer>
+    </S.Container>
+  );
+};
+
+export default Footer;
